Create README.md
this tools allows you to search on any username, name and surname. and search the results in two search engines and all over the web! when it intercepts your request and finds the username or name and surname you are looking for, it will print the urls of the page it is on on the terminal! and much more.! great for making osint 

commands

cd intercept

pip3 install requirements.txt

python3 intercept.py